make core as a globle module
